Replaces all references to GroupUser with Scholarship
