+++ app/models/user.rb
@@ -9,8 +9,6 @@ class User < ApplicationRecord
   devise :database_authenticatable, :registerable,
     :recoverable, :rememberable, :validatable
 
-  has_many :groups_users
-  has_many :groups, through: :groups_users
   has_one :student
 
   accepts_nested_attributes_for :student
