--- app/views/groups_users/new.html.erb
@@ -1,15 +1,0 @@
-
-         <main style="padding-left: 240px">
-
-            <div class="row">
-          <div class="col s12" style="padding-top: 20px;"><h4 class="center-align">Nueva Convocatoria</h4> </div>
-            <div class="col s12 m4 l2"><p></p></div>
-          <div id="man" class="col s12 m4 l8">
-			<%= render "form", group_user: @groups_user, groups: @groups %>
-          </div>
-          <div class="col s12 m4 l2"><p></p></div> 
-        </div>
-
-
-
-          </main>
