+++ app/views/scholarships/_form.html.erb
@@ -1,0 +1,5 @@
+<%= form_tag import_scholarships_path, multipart: true do %>
+  <%= select_tag 'scholarship_oportunity_id', options_for_select(scholarship.collect{ |u| [u.name, u.id] }) , class: "browser-default"%>
+  <%= file_field_tag :file,:style=>'padding-top:10px', class: 'form-control' %><br>
+  <%= submit_tag 'Import',:style=>'color:white;font-size:10px;width:110px ! important', class: 'btn btn-primary'%>
+<% end %>
