+++ config/routes.rb
@@ -15,8 +15,7 @@ Rails.application.routes.draw do
   post 'import_users' => 'users#import_user', as: 'import_users'
   resources :students
   resources :pages
-  resources :groups 
-  resources :groups_users
-  post 'import_groups_users' => 'groups_users#import_groups_users', as: 'import_groups_users'
-  # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
+  resources :scholarships
+  resources :scholarship_oportunities
+  post 'import_scholarships' => 'scholarships#import_scholarships', as: 'import_scholarships'
 end
